fix color scheme regression for highlights
